Ability to perform operations in sql through sts

+++ src/main/java/com/example/addressbook/repository/AddressBookRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.addressbook.model.AddressBookData;
 
-public interface AddressBookRepository extends JpaRepository<AddressBookData,Long> {
+public interface AddressBookRepository extends JpaRepository<AddressBookData,Integer> {
 
 }
